Create plot2.R
R code for plot 2. The plot itself may be viewed after clicking on https://github.com/zubr/ExData_Plotting1 
then click on "... commits" and the plot is attached as a comment to commit ed35c057f529643ea72e6df1fc620655851752ff
Sorry for that but I cannot install github on my pc (windows XP) and there seem to be no other option to download a chart on github (i.e. without cloning the repo in desktop).